requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
Django==3.2
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
idna==2.10
oauthlib==3.1.1
phonenumbers==8.12.21
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
